Feat : 캐시 로그

+++ 2Week_Mission/src/main/java/com/ll/exam/final__2022_10_08/app/cash/entity/CashLog.java
@@ -18,7 +18,7 @@ import static javax.persistence.FetchType.LAZY;
 @SuperBuilder
 @ToString(callSuper = true)
 public class CashLog extends BaseEntity {
-    private int price; // 변동가격
+    private long price; // 변동가격
     private String eventType; // 변동 종류
 
     @ManyToOne(fetch = LAZY)
